chore(deps): drop @mastra/* + @powerhousedao/ph-clint + codegen runtime dep
ph-cli:
- Delete src/code/ (the `ph code` Mastra REPL command). It was the only
  consumer of `@mastra/core`, `@mastra/libsql`, `@mastra/memory`, and
  `@powerhousedao/ph-clint`; drop all four.
- Drop `code` from `phCliCommands` registration.
- Drop `code` from `phCliCommandNames`, `phCliHelpCommands`, and the
  `args/code.ts` definition + re-exports in `@powerhousedao/shared`.

ph-cmd:
- Move `commands/checkout.ts` and its remote-drive validators to a
  top-level `legacy/` folder (excluded from tsconfig). Checkout was already
  not registered in `commands/ph.ts`, so it was dead in the published
  binary already; the move makes it explicit and removes the codegen +
  common imports from the active source tree.
- Refactor `commands/setup-globals.ts` to reuse `ph init` instead of
  calling `createProject` from `@powerhousedao/codegen` directly. Both
  `init.ts` and `setup-globals.ts` now share `utils/delegate-init.ts`
  which dlx-delegates to `@powerhousedao/ph-cli init` (with `--name .ph`
  injected for setup-globals, plus a chdir to HOME and a post-step that
  renames the generated `package.json#name` away from the invalid `.ph`).
- Move `generate-commands-docs.ts` from `src/` to `scripts/` and run it
  via `bun run` instead of bundling. Bundling broke after codegen became
  a devDep (chalk's CJS interop failed under tsdown). Mirrors ph-cli's
  pattern.
- Add `src/get-version.ts` (mirrors ph-cli) so the version string falls
  back to env vars when the source is run directly via bun (postbuild +
  `pnpm dev`), instead of relying on tsdown's `define` injection.
- Delete `src/utils/{constants,dependencies,index,package-manager,types}.ts`
  — none of them were imported by live source after the legacy move.
- Drop `@powerhousedao/codegen` and `@powerhousedao/common` from
  `dependencies`; keep `@powerhousedao/codegen` as a devDep for the
  build-time docs generator.

Net delta: ~250 fewer transitive packages installed (the @mastra/* +
@ai-sdk/* tree alone is sizable).
