+++ clis/ph-cli/package.json
@@ -38,14 +38,10 @@
     "@powerhousedao/codegen": "workspace:*",
     "@powerhousedao/common": "workspace:*",
     "@powerhousedao/config": "workspace:*",
-    "@powerhousedao/ph-clint": "^0.1.0-dev.21",
     "@powerhousedao/reactor": "workspace:*",
     "@powerhousedao/shared": "workspace:*",
     "@powerhousedao/switchboard": "workspace:*",
     "@renown/sdk": "workspace:*",
-    "@mastra/core": "^1.22.0",
-    "@mastra/libsql": "^1.7.4",
-    "@mastra/memory": "^1.13.1",
     "chalk": "catalog:",
     "change-case": "catalog:",
     "cmd-ts": "catalog:",
