+++ clis/ph-cli/src/commands/ph-cli-commands.ts
@@ -1,6 +1,5 @@
 import { accessToken } from "./access-token.js";
 import { build } from "./build.js";
-import { code } from "./code.js";
 import { connect } from "./connect.js";
 import { generate } from "./generate.js";
 import { init } from "./init.js";
@@ -19,7 +18,6 @@ import { vetra } from "./vetra.js";
 
 export const phCliCommands = {
   init,
-  code,
   generate,
   vetra,
   connect,
