+++ clis/ph-cmd/legacy/utils/create-global-project.ts
