+++ clis/ph-cmd/legacy/utils/validate-remote-drive-checkout.ts
