+++ clis/ph-cmd/legacy/utils/validate-remote-drive.ts
