+++ clis/ph-cmd/package.json
@@ -23,7 +23,7 @@
   "scripts": {
     "tsc": "tsc",
     "lint": "eslint",
-    "generate-docs": "node ./dist/generate-commands-docs.mjs",
+    "generate-docs": "bun run scripts/generate-commands-docs.ts",
     "build": "tsdown",
     "postbuild": "bun generate-docs",
     "dev": "bun run ./src/cli.ts",
@@ -34,19 +34,17 @@
   "keywords": [],
   "author": "",
   "dependencies": {
-    "@powerhousedao/codegen": "workspace:*",
-    "@powerhousedao/common": "workspace:*",
     "@powerhousedao/shared": "workspace:*",
     "chalk": "catalog:",
     "cmd-ts": "catalog:",
     "package-manager-detector": "catalog:",
     "read-pkg": "catalog:",
-    "write-package": "catalog:",
-    "semver": "catalog:"
+    "semver": "catalog:",
+    "write-package": "catalog:"
   },
   "devDependencies": {
+    "@powerhousedao/codegen": "workspace:*",
     "tsdown": "catalog:",
-    "vitest": "catalog:",
-    "commander": "catalog:"
+    "vitest": "catalog:"
   }
 }
