+++ clis/ph-cmd/scripts/generate-commands-docs.ts
@@ -1,10 +1,10 @@
 import { writeCliDocsMarkdownFile } from "@powerhousedao/codegen/file-builders";
-import { init } from "./commands/init.js";
-import { ph } from "./commands/ph.js";
-import { setupGlobals } from "./commands/setup-globals.js";
-import { update } from "./commands/update.js";
-import { useLocal } from "./commands/use-local.js";
-import { use } from "./commands/use.js";
+import { init } from "../src/commands/init.js";
+import { ph } from "../src/commands/ph.js";
+import { setupGlobals } from "../src/commands/setup-globals.js";
+import { update } from "../src/commands/update.js";
+import { useLocal } from "../src/commands/use-local.js";
+import { use } from "../src/commands/use.js";
 
 const commands = [
   { name: "init", command: init },
@@ -15,12 +15,13 @@ const commands = [
 ];
 
 const cliDescription = ph.description ?? "";
-declare const CLI_VERSION: string;
+const version =
+  process.env.WORKSPACE_VERSION ?? process.env.npm_package_version ?? "unknown";
 
 async function main() {
   await writeCliDocsMarkdownFile({
     filePath: "COMMANDS.md",
-    docsTitle: `PH-CMD CLI Commands (${CLI_VERSION})`,
+    docsTitle: `PH-CMD CLI Commands (${version})`,
     docsIntroduction:
       "This document provides detailed information about the available commands in the PH-CMD CLI. The CLI is published as `ph-cmd` and is invoked with the `ph` command.",
     cliDescription,
