--- clis/ph-cmd/src/utils/index.ts
@@ -1,4 +1,0 @@
-export * from "./constants.js";
-export * from "./dependencies.js";
-export * from "./package-manager.js";
-export type * from "./types.js";
