+++ packages/shared/clis/index.mts
@@ -1,5 +1,4 @@
 export * from "./args/access-token.js";
-export * from "./args/code.js";
 export * from "./args/common.js";
 export * from "./args/connect.js";
 export * from "./args/generate.js";
